+++ job.h
@@ -13,16 +13,16 @@ class job
 		jobType jobClass;
 		
 		// Time Required
-		float timeRequired;
+		double timeRequired;
 		
 		// Time it has been serviced so far
-		float timeServiced;
+		double timeServiced;
 		
 		// Arrival Time
-		float arrivalTime;
+		double arrivalTime;
 		
 		// Completion Time
-		float completionTime;
+		double completionTime;
 		
 		// Is this the last time this job needs to be serviced?
 		bool last;
@@ -33,7 +33,7 @@ class job
 		job();
 		
 		// Parameterized Constructor
-		job( int num, jobType type, float timeReq, float timeServd, float timeArrive, float timeCompleted, bool f );
+		job( int num, jobType type, double timeReq, double timeServd, double timeArrive, double timeCompleted, bool f );
 	
 	  //getters
 		// Return the Job Numer
@@ -43,16 +43,16 @@ class job
 		jobType getJobType() const;
 		
 		// Return the time this job requires
-		float getTimeRequired() const;
+		double getTimeRequired() const;
 		
 		//Return the time this job has been serviced thus far
-		float getTimeServiced() const;
+		double getTimeServiced() const;
 		
 		//Return the arrival time of this job
-		float getArrivalTime() const;
+		double getArrivalTime() const;
 		
 		// Return the completion time of this job
-		float getCompletionTime() const;
+		double getCompletionTime() const;
 		
 		// Return whether this is the first time this job has been serviced this timeslice
 		bool lastService() const;
@@ -65,19 +65,19 @@ class job
 	  	void setJobType( jobType t );
 	  	
 	  	//Set the amount time required for this job
-	  	void setTimeRequired( float t );
+	  	void setTimeRequired( double t );
 	  	
 	  	// Set the amount of time this job has been serviced thus far
-	  	void setTimeServiced( float t );
+	  	void setTimeServiced( double t );
 	  	
 	  	// Increment the time this job has been serviced by .1
 	  	void incrementTimeServiced();
 	  	
 	  	// Set the arrival time of this job
-	  	void setArrivalTime( float t );
+	  	void setArrivalTime( double t );
 	  	
 	  	// Set the Completion time of this job
-	  	void setCompletionTime( float t );
+	  	void setCompletionTime( double t );
 	  	
 	  	// Set whether this is the first time this job had been serviced this timeslice
 	  	void setLastService( bool l );
